requirements.txt
# Core dependencies
fastapi 
uvicorn
python-multipart
pydantic

# LangChain and related
langchain
langchain-community
langchain-core
langchain-openai
langchain-chroma
chromadb

# Document processing
pytesseract
pdf2image
Pillow
pypdf
graphviz

# Environment and utilities
python-dotenv
openai
certifi
urllib3

# Optional but recommended
numpy
pandas

watchfiles